Bitstreams

All bitstreams inherit from a shared abstract type, AbstractBitstream. This type defines a bitstream as a queue (sequence) of bits. Since the underlying "bit" in BitSAD depends on the type of bitstream being used, we also define an abstract bit type, AbstractBit.

BitSAD.AbstractBit — Type

AbstractBit

Inherit from this type to create a custom bit type.

BitSAD.AbstractBitstream — Type

AbstractBitstream

Inherit from this type to create a custom bitstream type.

Expected fields:

bits::Queue{AbstractBit}: the underlying bitstream

Any bitstream has several common operations defined on it.

Base.push! — Function

push!(s::AbstractBitstream, b)

Push a bit(s) b onto bitstream s.

Fields:

s::AbstractBitstream: the bitstream object

b: the bit(s) to push onto the stream

Base.pop! — Function

pop!(s::AbstractBitstream)

Pop a bit from bitstream s.

Fields:

s::AbstractBitstream: the bitstream object

BitSAD.observe — Function

observe(s::AbstractBitstream)

Examine the most recent bit added to the stream without removing it.

Fields:

s::AbstractBitstream: the bitstream object

Base.length — Function

length(s::AbstractBitstream)

Return the number of bits in s.