Hardware internals

From the user perspective, generating hardware seems fairly trivial. Yet, there is a lot happening under the hood to make this possible. It will be important to understand these details if you want to customize the hardware generation process.

All the information for hardware generation is stored in a Module object. The HW.generate function operates on this object to output a Verilog string. Users will typically use @circuit to generate the module object.

BitSAD.HW.Module — Type

Module

A data structure to store information to generate hardware for a circuit. This structure can be manually modified if needed but typically @circuit is used to auto-populate it.

Hardware generation traverses dfg and uses handlers to generate Verilog strings.

Fields:

name::Symbol: the name of the module

parameters::Dict{Symbol, Number}: a map from the name of each parameter to its default value

submodules::Dict{Symbol, Symbol}: a map from the name of each submodule to its type

dfg::MetaDiGraph{Int, Float64}: a data flow graph representing the circuit to be generated

handlers::Dict{Operation, AbstractHandler}: a map from operation type to a hardware generation handler.

See also: HW.generate

BitSAD.HW.generate — Function

HW.generate(m::Module, netlist::Netlist)
HW.generate(m::Module, f)
HW.generate(c::Tuple{Module, Function}, dut, args...)

Generate the Verilog implementation of the module. Users will most likely call the last method form above.

Fields:

m::Module: the module to generate

netlist::Netlist: the netlist for the circuit being generated

f: a closure with one argument (a netlist) that calls a runtime information extraction function

c::Tuple{Module, Function}: the tuple returned by @circuit

dut: an instance of the circuit struct

args: example arguments to circuit

BitSAD.HW.@circuit — Macro

@circuit

Create a new BitSAD circuit by specifying the name, parameters, submodules, initialization, and implementation algorithm.

Returns a Tuple{Module, Function}. See Module. The function returned extracts runtime information to store in the Module.

The function has the signature where T is the name of the circuit generated:

extractrtinfo!(dut::T, m::BitSAD.HW.Module, netlist::BitSAD.HW.Netlist, args...)

Examples

m = @circuit Foo begin
    parameters : [
        a => 0.125
        b => 10
    ]

    submodules : [
        m::SDM,
        n::CircularBuffer{DBit}
    ]

    initialize : begin
        dut = new(a, b, m, n)
        fill!(dut.n, zero(DBit))

        return dut
    end

    circuit : (dut::Foo)(x::DBit) -> begin
        y = dut.a * x
        z = y + dut.b

        return z
    end
end

@circuit internals

The @circuit is responsible for parsing a domain-specific language for creating and populating Module objects. In Generating hardware, we see an example of @circuit being used. Here is an equivalent piece of code for achieving the same goal without using @circuit

using BitSAD
using BitSAD.HW: Variable

Base.@kwdef struct IterativeSVD
    rows::Int = 2
    cols::Int = 2
end

function (dut::IterativeSVD)(A::Matrix{SBit}, v₀::Vector{SBit})
    # Update right singular vector
    w = A * v₀
    wscaled = w .÷ sqrt(dut.rows)
    u = wscaled ./ norm(wscaled)

    # Update left singular vector
    z = permutedims(A) * u
    zscaled = z .÷ sqrt(dut.cols)
    σ = norm(zscaled)
    v = zscaled ./ σ

    return u, v, σ
end

# manually create Module
m = Module(name = Symbol("IterativeSVD"))

# manually populate module DFG
inputs = [Variable(:A, :Any), Variable(:v₀, :Any)]
outputs = [Variable(:w, :Any)]
op = :*
addnode!(m, inputs, outputs)
# ... remaining addnode! calls left out for brevity

# define runtime info extraction function
function extractrtinfo!(dut::IterativeSVD, m::BitSAD.HW.Module, netlist::BitSAD.HW.Netlist, A::Matrix{SBit}, v₀::Vector{SBit})
    w = extractrtinfo!(m, netlist, [:A, :v₀], :w, *, false, nothing, A, v₀)
    wscaled = extractrtinfo!(m, netlist, [:w, :net_2_2_1], :wscaled, div, true, nothing, w,
                    extractrtinfo!(m, netlist, [:rows], :net_2_2_1, sqrt, false, nothing, dut.rows)
                )
    # ... remaining lines ommitted for brevity
end

circuit = (m, extractrtinfo!)

We ommitted several lines to keep the snippet above short, but as you can see, for each operation in the DFG, there are many auto-generated lines of code.

HW.generate internals

The HW.generate goes through several stages to produce the output Verilog string. Here are those steps (assuming circuit is the tuple returned by @circuit):

Call HW.generate(circuit, dut, args...) where dut refers to an instance of the algorithm struct and args... are the example arguments to the circuit.

This function in turn calls HW.generate(circuit[1], (netlist) -> circuit[2](dut, circuit[1], netlist, args...)).

This function in turn does:

Creates a new netlist with netlist = Netlist().

Populates the module and netlist with runtime information calling f(netlist) where f is the anonymous function in Step 2 above.

Calls HW.generate(module, netlist) where module is circuit[1] populated with runtime info using f.

We now step through the main hardware generation function:

Apply the constant reduction phase to reduce expressions containing only compile time literals or module parameters.

Apply the constant replacement phase to transform each constant in the circuit into a binary Verilog literal.

Traverse the DFG from inputs to outputs in breadth-first fashion. For each node in the traversal:

Get the current handler object corresponding to the operation and input/output types of the node.

Call the handler object passing the current node's inputs/outputs and netlist. The handler will update the netlist if necessary and return a Verilog string instantiating that operation into the total circuit.

Return all the Verilog strings concatenated together.

HW.generate makes a deep-copy of the Module object, so any optimization phases above do not modify the DFG generated by @circuit. This allows you to call HW.generate multiple times with different example arguments. This can be useful if you want to programmatically generate many versions of the same circuit with different parameters (e.g. different matrix sizes).