Stochastic Bitstreams

An SBitstream is a stochastic bitstream representing a sequence of SBits. Each SBitstream is associated with an underlying real value.

BitSAD.SBit — Type

SBit

A stochastic bit is a pair of unipolar bits (positive and negative channels).

Fields:

bit::Tuple{Bool, Bool}: a sample of a bitstream

value::Float64: the underlying floating-point number being represented

id::UInt32: a unique identifier for all samples of this bitstream

BitSAD.SBitstream — Type

SBitstream

A stochastic bitstream that represents a real (floating-point) number between [-1, 1].

Fields:

bits::Queue{SBit}: the underlying bitstream

value::Float64: the underlying floating-point number being represented

id::UInt32: a unique identifier for this bitstream (set automatically)

To represent signed numbers in $[-1, 1]$, we use two single bitstreams — a positive channel and a negative channel. So, an SBit is actually a tuple of two boolean values.

BitSAD.pos — Function

pos(b::SBit)

Return the positive channel bit of a stochastic bit.

BitSAD.neg — Function

neg(b::SBit)

Return the negative channel bit of a stochastic bit.

We can also access the underlying real value using float.

Base.float — Method

float(b::SBit)
float(s::SBitstream)

Return the underlying floating-point value of a stochastic bit or bitstream.

Finally, we can fill up a SBitstream with a bit sequence using generate! and estimate the empirical average using estimate!.

BitSAD.generate — Function

generate(s::SBitstream, T::Integer = 1)
generate!(s::SBitstream, T::Integer = 1)

Generate T samples of the bitstream. Add them to its queue for generate!.

BitSAD.estimate! — Function

estimate!(buffer::AbstractVector, b::SBit)
estimate!(buffer::AbstractVector, b::VecOrMat{SBit})
estimate!(buffer::AbstractVector)

Push b into the buffer and return the current estimate.

Operators

The following operations are defined for SBits and SBitstreams.

Operation	Name	Conditions

+(x::SBit, y::SBit)	Addition	None

-(x::SBit, y::SBit)	Subtraction	None

*(x::SBit, y::SBit)	Multiplication	None

/(x::SBit, y::SBit)	Division	y > 0

÷(x::SBit, y::Real)	Fixed-Gain Division	y >= 1

sqrt(x::SBit)	Square Root	x >= 0

norm(x::Vector{SBit})	L2 Norm	None